Update ard-freezercontrol.ino
The important change "on/off-regulation" instead of PID regulation.

Inn addition initial settpoint is changed to +4 and it is added a lot off comments.

Future work:
Measure % time the relay is on and off
Measure time it take for the from the realy is tuned on until it is tuned on again. This delta time should be shown on the LCD instead of "Output". If the time is too short the hysterese must be increased.